+++ lib_dolphin/htk_helpers.py
@@ -1,7 +1,13 @@
 import numpy as np
 import struct
 import pandas as pd
+import sys
+import glob
 
+from subprocess import check_output
+
+
+FLOOR = 1.0
 PERIOD      = 1
 SAMPLE_SIZE = 4 
 USER        = 9
@@ -125,3 +131,83 @@ def mmf(label_file, proto_file, hmm_out="hmm0", hmm_list_out="monophones"):
 
 def htk_name(num):
     return "".join([str(chr(int(i) + 97)) for i in str(num)])
+    
+
+def get_ll(out):
+    out = out.split(b"\n")
+    for line in out:
+        if b"average log prob per frame" in line:
+            ll = line.strip().split(b" ")[-1]
+            ll = float(ll)
+            return ll
+
+        
+def take_step(folder, i):
+    files = glob.glob("{}/data/train/*.htk".format(folder))
+    out = check_output(["rm", "-rf", "{}/hmm{}".format(folder, i)])
+    out = check_output(["mkdir", "{}/hmm{}".format(folder, i)])
+    out = check_output("HERest -A -T 1 -v {} -I {}/clusters_TRAIN.mlf -M {}/hmm{} -H {}/hmm{}/hmm_mmf {}/list".format(FLOOR, folder, folder, i, folder, i - 1, folder).split(" ") + files)
+    return get_ll(out)
+
+
+def htk_eval(folder, last_hmm):
+    files = glob.glob("{}/data/test/*.htk".format(folder))
+    out = check_output("HVite -T 1 -n 10 -p 0.0 -s 5.0 -H {}/hmm{}/hmm_mmf -i {}/predictions.mlf -w {}/wdnet {}/dict {}/list".format(
+        folder, last_hmm, folder, folder, folder, folder
+    ).split(" ") + files)
+    out = check_output("HResults -I {}/clusters_TEST.mlf {}/list {}/predictions.mlf".format(folder, folder, folder).split(" "))
+    return out.decode("utf-8")
+    
+
+def htk_export(folder, out_htk, out_lab, k=46, min_c = 10):
+    instances_file   = "{}/instances.pkl".format(folder)
+    predictions_file = "{}/predictions.pkl".format(folder)
+    clusters_file    = "{}/clusters.pkl".format(folder)
+    label_file       = "{}/labels.pkl".format(folder)
+    
+    instances   = pkl.load(open(instances_file, "rb")) 
+    clusters    = pkl.load(open(clusters_file, "rb"))[k, :]
+    predictions = pkl.load(open(predictions_file, "rb")) 
+    label_dict  = pkl.load(open(label_file, "rb"))
+    reverse     = dict([(v,k) for k, v in label_dict.items()])
+    
+    ids_cluster = {}
+    for i, cluster in enumerate(clusters):
+        if cluster not in ids_cluster:
+            ids_cluster[cluster] = []
+        ids_cluster[cluster].append(i)
+        
+    cur = 0
+    label_dict = {}
+    train = "{}_TRAIN.mlf".format(out_lab.replace(".mlf", ""))
+    test  = "{}_TEST.mlf".format(out_lab.replace(".mlf", ""))
+    os.system("mkdir {}/train".format(out_htk))
+    os.system("mkdir {}/test".format(out_htk))
+    n_exp = 0
+    with open(train, 'w') as fp_train, open(test, 'w') as fp_test:
+        fp_train.write("#!MLF!#\n")
+        fp_test.write("#!MLF!#\n")
+        for c, ids in ids_cluster.items():
+            label = label_cluster(predictions, ids, reverse)
+            if c not in label_dict:
+                label_dict[c] = htk_name(c)
+            if label != "ECHO" and len(ids) > min_c:
+                n_exp += 1
+                random.shuffle(ids)                
+                train_ids = ids[0:len(ids)//2]
+                test_ids  = ids[len(ids)//2:len(ids)]
+                for i in train_ids:
+                    n = len(instances[i])
+                    write_htk(instances[i], "{}/train/{}_{}.htk".format(out_htk, label_dict[c], i))
+                    fp_train.write("\"*/{}_{}.lab\"\n".format(label_dict[c], i))
+                    fp_train.write("{} {} {}\n".format(0, n, label_dict[c]))
+                    fp_train.write(".\n")
+                for i in test_ids:
+                    n = len(instances[i])
+                    write_htk(instances[i], "{}/test/{}_{}.htk".format(out_htk, label_dict[c], i))
+                    fp_test.write("\"*/{}_{}.lab\"\n".format(label_dict[c], i))
+                    fp_test.write("{} {} {}\n".format(0, n, label_dict[c]))
+                    fp_test.write(".\n")                    
+    for c, k in label_dict.items():
+        print("{}\t{}".format(c, k))
+    print("#clusters: {}".format(n_exp))
